Add "print entire section" feature
It's sometimes useful to be able to create a printout of all the pages
within a documentation or blog section, without having to select each
one at a time.

This commit adds a "print this section" button to the right navigation
by creating a new "print" output format and applying it to sections.

Each sub-section starts a new page.

Partials can be created to customize the layout on a per-type basis for:

* print/page-heading-<type>.html
* print/toc-li-<type>.html
* print/content-<type>.html

Pages can be omitted from a printed section entirely by setting
no_print=true in the page front matter.

Types can be excluded from being printable by setting outputs explicitly
in their front matter and listing only HTML, RSS, etc (in conjunction
with cascade)

To enable the feature, add the following to the site's config.toml

[outputs]
section = [ "HTML", "RSS", "print"]